+++ src/lib/chat/slots/MessageTimeDefault.svelte
@@ -3,6 +3,7 @@
   import { faCheck, faCheckDouble, faTriangleExclamation, faComment, faClock } from '@fortawesome/free-solid-svg-icons';
   import { faWhatsapp, faTelegram } from '@fortawesome/free-brands-svg-icons';
   import Icon from '../../ui/display/Icon.svelte';
+  import MessengerIcon from './MessengerIcon.svelte';
   import type { Message } from '../types';
   import { deliveryTick } from '../helpers';
 
@@ -38,5 +39,8 @@
         <Icon icon={channelIcon(ch.channel)} size="xs" />
       </span>
     {/each}
+  {/if}
+  {#if !isMine}
+    <MessengerIcon {message} class="text-[10px]" />
   {/if}
 </span>